+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `--path-style` with `base-dir` (default), `canonical`, `dirname`, `short` and `working-dir`. Replacement for `--canonical` and `--relative`.
 - Add options for command output handling:
   - `--no-header`
-  - `--output` with `inherit`, `line-prefix` and `null`
+  - `--output` with `inherit`, `line-prefix`, `collect` and `null`
   - `--result` with `always`, `never` and `non-zero`
   - `--line-prefix` as a shortcut for `--no-header --output=line-prefix`
   - `--only-result` as a shortcut for `--no-header --output=null`